🔄 Major refactor for consistency and transparency
📂 **Directory Structure**:
  - Renamed build/ → build-linux/ (clearer purpose)
  - Restored .gitignore to ignore build/ (for electron-builder output)
  - Added platform-specific package.json files to .gitignore

⚡ **Package.json Scripts Refactored**:
  - package:setup → linux:setup
  - package:linux → linux:all
  - package:flatpak → linux:flatpak
  - package:appimage → linux:appimage
  - package:deb → linux:deb
  - package:rpm → linux:rpm
  - package:test → linux:test
  - package:restore → linux:restore

🔧 **Script Improvements**:
  - Fixed duplication in build-all.ts - now properly uses handle-platform-deps.ts
  - Updated all TypeScript scripts to use build-linux/ paths
  - Added clear comment to tsconfig.json (scripts-only usage)
  - Updated README.md with new command structure

✅ **Tested and Working**:
  - npm run linux:setup ✅
  - npm run linux:test ✅ (builds Docker images correctly)
  - npm run linux:restore ✅
  - All path references updated correctly

🎯 **Benefits**:
  - Clear separation: build-linux/ for packaging, build/ for electron-builder
  - Consistent linux:* naming for all Linux packaging commands
  - No more duplication between build scripts
  - Better transparency about what each command does
