+++ .gitignore
@@ -181,14 +181,17 @@ temp/
 # Electron specific
 out/
 dist/
-# Note: build/ directory contains packaging scripts and manifests, not build output
-# Build output goes to dist/. Only ignore build/Release from native modules
+build/
 *.dmg
 *.exe
 *.deb
 *.rpm
 *.AppImage
 *.snap
+
+# Platform-specific package.json files (generated by build scripts)
+package.*.json
+package.json.backup
 
 # macOS
 .DS_Store
