+++ build-linux/rpm/open-wispr.spec
