+++ package.json
@@ -21,14 +21,14 @@
     "lint": "cd src && eslint .",
     "preview": "cd src && vite preview",
     "clean": "node cleanup.js",
-    "package:setup": "npx ts-node build/scripts/handle-platform-deps.ts create-platform-configs",
-    "package:linux": "npx ts-node build/scripts/build-all.ts",
-    "package:flatpak": "npx ts-node build/scripts/build-flatpak.ts",
-    "package:appimage": "npx ts-node build/scripts/build-appimage.ts",
-    "package:deb": "npx ts-node build/scripts/build-deb.ts",
-    "package:rpm": "npx ts-node build/scripts/build-rpm.ts",
-    "package:test": "npx ts-node build/scripts/test-packages.ts",
-    "package:restore": "npx ts-node build/scripts/handle-platform-deps.ts restore"
+    "linux:setup": "npx ts-node build-linux/scripts/handle-platform-deps.ts create-platform-configs",
+    "linux:all": "npx ts-node build-linux/scripts/build-all.ts",
+    "linux:flatpak": "npx ts-node build-linux/scripts/build-flatpak.ts",
+    "linux:appimage": "npx ts-node build-linux/scripts/build-appimage.ts",
+    "linux:deb": "npx ts-node build-linux/scripts/build-deb.ts",
+    "linux:rpm": "npx ts-node build-linux/scripts/build-rpm.ts",
+    "linux:test": "npx ts-node build-linux/scripts/test-packages.ts",
+    "linux:restore": "npx ts-node build-linux/scripts/handle-platform-deps.ts restore"
   },
   "keywords": [
     "dictation",
